requirements.txt
torch
torchvision
numpy
scikit-learn
matplotlib
flask
pillow
